requirements.txt
# Core dependencies
ruamel.yaml>=0.17.0
pydantic>=2.0.0
click>=8.0
jsonschema>=4.0
jinja2>=3.0
